+++ src/ua/data_types.rs
@@ -11,7 +11,6 @@ mod browse_next_response;
 mod browse_request;
 mod browse_response;
 mod browse_result;
-mod browse_result_mask;
 mod byte_string;
 mod call_method_request;
 mod call_method_result;
@@ -56,8 +55,7 @@ pub use self::{
     argument::Argument, attribute_id::AttributeId, browse_description::BrowseDescription,
     browse_direction::BrowseDirection, browse_next_request::BrowseNextRequest,
     browse_next_response::BrowseNextResponse, browse_request::BrowseRequest,
-    browse_response::BrowseResponse, browse_result::BrowseResult,
-    browse_result_mask::BrowseResultMask, byte_string::ByteString,
+    browse_response::BrowseResponse, browse_result::BrowseResult, byte_string::ByteString,
     call_method_request::CallMethodRequest, call_method_result::CallMethodResult,
     call_request::CallRequest, call_response::CallResponse,
     create_monitored_items_request::CreateMonitoredItemsRequest,