+++ src/ua/data_types/attribute_id.rs
@@ -1,7 +1,5 @@
 use std::hash;
 
-use open62541_sys::UA_AttributeId;
-
 crate::data_type!(AttributeId, UInt32);
 
 crate::enum_variants!(
@@ -37,14 +35,6 @@ crate::enum_variants!(
         ACCESSLEVELEX,
     ],
 );
-
-impl AttributeId {
-    #[deprecated(note = "use `Self::VALUE` instead")]
-    #[must_use]
-    pub const fn value() -> Self {
-        Self(UA_AttributeId::UA_ATTRIBUTEID_VALUE)
-    }
-}
 
 impl hash::Hash for AttributeId {
     fn hash<H: hash::Hasher>(&self, state: &mut H) {