+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project
 adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Add logical OR combinator for `ua::BrowseResultMask` and `ua::NodeClassMask`.
+- Add const `ua::NodeClassMask` variants to initialize masks.
+
+### Changed
+
+- Breaking: Remove associated functions for enum data types deprecated in 0.5.0, e.g.
+  `ua::AttributedId::value()`. Use uppercase constants `ua::AttributedId::VALUE` instead.
+
 ## [0.6.0-pre.2] - 2024-04-12
 
 [0.6.0-pre.2]: https://github.com/HMIProject/open62541/compare/v0.6.0-pre.1...v0.6.0-pre.2